+++ app/models/spec_views/view_sanitizer.rb
@@ -1,0 +1,20 @@
+module SpecViews
+  class ViewSanitizer
+    delegate :each, to: :@sanitizers
+
+    def initialize
+      @sanitizers = []
+    end
+
+    def gsub(*args, &block)
+      @sanitizers << [:gsub, args, block]
+    end
+
+    def sanitize(string)
+      each do |sanitize_method, args, block|
+        string = string.send(sanitize_method, *args, &block)
+      end
+      string
+    end
+  end
+end